+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ Changelog
 * Update :meth:`pymaf.MafFrame.to_vcf` method to output sorted VCF.
 * Add :meth:`pymaf.MafFrame.matrix_prevalence` method.
 * Add :meth:`pymaf.MafFrame.plot_regplot` method.
+* Add ``samples`` argument to :meth:`pymaf.MafFrame.plot_snvclss` method.
 
 0.16.0 (2021-07-02)
 -------------------